+++ ichnaea/api/locate/internal.py
@@ -12,26 +12,34 @@ from ichnaea.api.locate.source import (
     PositionSource,
     RegionSource,
 )
-from ichnaea.api.locate.wifi import WifiPositionMixin
+from ichnaea.api.locate.wifi import (
+    WifiPositionMixin,
+    WifiRegionMixin,
+)
 
 
-class InternalRegionSource(CellRegionMixin, RegionSource):
+class InternalRegionSource(CellRegionMixin,
+                           WifiRegionMixin, RegionSource):
     """A region source based on our own crowd-sourced internal data."""
 
+    fallback_field = None  #:
     source = DataSource.internal  #:
 
     def should_search(self, query, results):
         if not RegionSource.should_search(
                 self, query, results):  # pragma: no cover
             return False
-        if not self.should_search_cell(query, results):
+        if not (self.should_search_cell(query, results) or
+                self.should_search_wifi(query, results)):
             return False
         return True
 
     def search(self, query):
         results = RegionResultList()
         for should, search in (
-                (self.should_search_cell, self.search_cell), ):
+            # start with cell search, we don't need precise results
+                (self.should_search_cell, self.search_cell),
+                (self.should_search_wifi, self.search_wifi)):
 
             if should(query, results):
                 results.add(search(query))
@@ -66,6 +74,7 @@ class InternalPositionSource(CellPositionMixin,
     def search(self, query):
         results = self.result_type().as_list()
         for should, search in (
+            # start with wifi search, we want precise results
                 (self.should_search_wifi, self.search_wifi),
                 (self.should_search_cell, self.search_cell)):